microservicios/indicadores_service/requirements.txt
flask
requests
pulsar-client